### 2023-03-31 92de074
adding interface to scipy
