+++ iwopy/wrappers/discretize_reg_grid.py
@@ -186,7 +186,7 @@ class DiscretizeRegGrid(LocalFD):
                 f"Problem '{self.name}' cannot apply non-grid points to problem"
             )
 
-    def evaluate_individual(self, vars_int, vars_float):
+    def evaluate_individual(self, vars_int, vars_float, ret_prob_res=False):
         """
         Evaluate a single individual of the problem.
 
@@ -196,6 +196,8 @@ class DiscretizeRegGrid(LocalFD):
             The integer variable values, shape: (n_vars_int,)
         vars_float : np.array
             The float variable values, shape: (n_vars_float,)
+        ret_prob_res : bool
+            Flag for additionally returning of problem results
 
         Returns
         -------
@@ -203,11 +205,13 @@ class DiscretizeRegGrid(LocalFD):
             The objective function values, shape: (n_objectives,)
         con : np.array
             The constraints values, shape: (n_constraints,)
+        prob_res : object, optional
+            The problem results
 
         """
         varsf = vars_float[self._vinds]
         if self.grid.is_gridpoint(varsf):
-            return super().evaluate_individual(vars_int, vars_float)
+            return super().evaluate_individual(vars_int, vars_float, ret_prob_res)
 
         else:
             gpts, coeffs = self.grid.interpolation_coeffs_point(varsf)
@@ -216,16 +220,30 @@ class DiscretizeRegGrid(LocalFD):
             objs = np.zeros((n_gpts, self.n_objectives), dtype=np.float64)
             cons = np.zeros((n_gpts, self.n_constraints), dtype=np.float64)
 
+            res = [None for __ in range(n_gpts)]
             for gi, gp in enumerate(gpts):
                 varsf = vars_float.copy()
                 varsf[self._vinds] = gp
-                objs[gi], cons[gi] = super().evaluate_individual(vars_int, varsf)
+                if ret_prob_res:
+                    objs[gi], cons[gi], res[gi] = super().evaluate_individual(
+                                                        vars_int, varsf, ret_prob_res)
+                else:
+                    objs[gi], cons[gi] = super().evaluate_individual(
+                                                    vars_int, varsf, ret_prob_res)
 
-            return np.einsum("go,g->o", objs, coeffs), np.einsum(
-                "gc,g->c", cons, coeffs
-            )
+            if ret_prob_res:
+                return (
+                    np.einsum("go,g->o", objs, coeffs), 
+                    np.einsum("gc,g->c", cons, coeffs),
+                    self.prob_res_einsum_individual(res, coeffs)
+                )
+            else:
+                return (
+                    np.einsum("go,g->o", objs, coeffs), 
+                    np.einsum("gc,g->c", cons, coeffs)
+                )
 
-    def evaluate_population(self, vars_int, vars_float):
+    def evaluate_population(self, vars_int, vars_float, ret_prob_res=False):
         """
         Evaluate all individuals of a population.
 
@@ -235,6 +253,8 @@ class DiscretizeRegGrid(LocalFD):
             The integer variable values, shape: (n_pop, n_vars_int)
         vars_float : np.array
             The float variable values, shape: (n_pop, n_vars_float)
+        ret_prob_res : bool
+            Flag for additionally returning of problem results
 
         Returns
         -------
@@ -242,13 +262,15 @@ class DiscretizeRegGrid(LocalFD):
             The objective function values, shape: (n_pop, n_objectives)
         cons : np.array
             The constraints values, shape: (n_pop, n_constraints)
+        prob_res : object, optional
+            The problem results
 
         """
         varsf = vars_float[:, self._vinds]
 
         # case all points on grid:
[…]
