+++ iwopy/interfaces/__init__.py
@@ -1,2 +1,3 @@
 from . import pygmo
 from . import pymoo
+from . import scipy
