+++ docs/source/history.md
@@ -113,3 +113,8 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
   - New interface to `scipy`: Calling the `scipy.optimize.minimize` function
   
 **Full Changelog**: [https://github.com/FraunhoferIWES/iwopy/commits/v0.1.6](https://github.com/FraunhoferIWES/iwopy/commits/v0.1.6)
+
+## v0.1.7-alpha
+
+  
+**Full Changelog**: [https://github.com/FraunhoferIWES/iwopy/commits/v0.1.6](https://github.com/FraunhoferIWES/iwopy/commits/v0.1.7)
