### 2023-04-17 795c45b
fix for issue #10 (github)
