+++ setup.cfg
@@ -48,6 +48,6 @@ doc =
     m2r2
 all =
     pygmo
-    pymoo
+    pymoo>=0.6
 
 scripts =
