### 2023-04-27 #5310 @arfon
:wave: @fraukewiese – this submission seems to be related to wind farm design optimization. Would you be willing to edit this for JOSS?
