### 2023-04-28 #5310 @fraukewiese
As far as I understand, it is not about a wind farm design optimization but rather about optimization in Python in general. However, I think I can edit this submission.
