+++ paper/paper.bib
@@ -1,11 +1,12 @@
-@ARTICLE{pymoo,
-    author={J. {Blank} and K. {Deb}},
-    journal={IEEE Access},
-    title={pymoo: Multi-Objective Optimization in Python},
-    year={2020},
-    volume={8},
-    number={},
-    pages={89497-89509},
+
+@article{pymoo,
+  title={Pymoo: Multi-Objective Optimization in Python},
+  author={Julian Blank and Kalyanmoy Deb},
+  journal={IEEE Access},
+  year={2020},
+  volume={8},
+  pages={89497-89509},
+  doi={10.1109/access.2020.2990567}
 }
 
 @article{pygmo,
