### 2023-05-08 #5310 @SchmJo
@fraukewiese I added the missing doi and pushed again to branch "paper".
