+++ README.md
@@ -38,6 +38,7 @@ The supported Python versions are:
 - `Python 3.8`
 - `Python 3.9`
 - `Python 3.10`
+- `Python 3.11`
 
 ## Installation via conda
 
