### 2023-08-21 #5310 @coroa
hi @fraukewiese , thanks for the invite. I am mostly unfamiliar with the optimization libraries this `iwopy` library aims at abstracting over: `pymoo` and `pygmo`.

It seems to me that their main use is for defining problem and solver/algorithm classes and then enabling to test either new algorithms on a set of already defined problems or defining new problem classes and then testing out a set of solvers on them.

I would rather not want to review given time-constraints and unfamiliarity.
