### 2023-09-25 #5310 @fraukewiese
@arjunane , @gaxler – would any of you be willing to review this submission https://github.com/FraunhoferIWES/iwopy  for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
