### 2023-10-06 #5310 @fraukewiese
@StewMH , @BenWinchester – would one of you be willing to review this submission https://github.com/FraunhoferIWES/iwopy for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
