### 2023-11-02 #6014 @editorialbot
[REVIEW]: IWOPY: Fraunhofer IWES optimization tools in Python
**Submitting author:** @SchmJo (<a href="http://orcid.org/0000-0002-8191-8141">Jonas Schmidt</a>)
**Repository:** https://github.com/FraunhoferIWES/iwopy
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.0.2
