### 2023-11-06 FraunhoferIWES/iwopy#11 @BenWinchester
Missing instructions for `tests`
**Describe the bug**
The automated tests contained within the `tests` directory cannot be easily run, potentially due to a lack of clear documentation or possibly an issue around imports as is suggests in the tracebacks when trying to run these tests.

**To Reproduce**
- Running with `pytest`
  - In an environment _without_ `iwopy` installed _or_ in an environment _with_ `iwopy` installed
    ```bash
    (python-3-10-environment) user@machine:~/path/to/iwopy/repository$ python -m pytest tests
    ================================================================== test session starts
    ==================================================================
    rootdir: /home/bewinche/modelling/reviewing/iwopy
    collected 0 items / 7 errors
    ......
    ======================================================================== ERRORS =========================================================================
    ______________________________________________________ ERROR collecting tests/scipy/test_scipy.py
    ImportError while importing test module '/home/bewinche/modelling/reviewing/iwopy/tests/scipy/test_scipy.py'.
    Traceback:
    ../../../anaconda3/envs/py310/lib/python3.10/importlib/__init__.py:126: in import_module
    tests/scipy/test_scipy.py:3: in <module>
        from iwopy import DiscretizeRegGrid, SimpleConstraint
    iwopy/__init__.py:12: in <module>
        from . import interfaces
    iwopy/interfaces/__init__.py:2: in <module>
        from . import pymoo
    iwopy/interfaces/pymoo/__init__.py:1: in <module>
        from .optimizer import Optimizer_pymoo, DefaultCallback
    iwopy/interfaces/pymoo/optimizer.py:5: in <module>
        from .problem import SingleObjProblem, MultiObjProblem
    iwopy/interfaces/pymoo/problem.py:4: in <module>
        from .imports import check_import, Problem, Real, Integer
    E   ImportError: cannot import name 'Real' from 'iwopy.interfaces.pymoo.imports' (/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/pymoo/imports.py)
    ================================================================= short test summary info ================================================================
    ERROR tests/core/test_discr_grid.py
    ERROR tests/core/test_discretize.py
    ERROR tests/core/test_gradients_ana.py
    ERROR tests/core/test_varmap.py
    ERROR tests/pygmo/test_pygmo.py
    ERROR tests/pymoo/test_pymoo.py
    ERROR tests/scipy/test_scipy.py
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 7 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    ==================================================================== 7 errors in 1.43s ===================================================================
      ```
- Running with Python directly
  - In an environment _without_ `iwopy` installed
    ```bash
    (python-3-10-without-iwopy) user@machine:~/path/to/iwopy/repository$ python tests/core/test_discr_grid.py
    Traceback (most recent call last):
      File "/home/bewinche/modelling/reviewing/iwopy/tests/core/test_discr_grid.py", line 3, in <module>
        import iwopy
      File "/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/__init__.py", line 12, in <module>
        from . import interfaces
      File "/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/interfaces/__init__.py", line 2, in <module>
        from . import pymoo
      File "/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/interfaces/pymoo/__init__.py", line 1, in <module>
        from .optimizer import Optimizer_pymoo, DefaultCallback
      File "/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/interfaces/pymoo/optimizer.py", line 5, in <module>
        from .problem import SingleObjProblem, MultiObjProblem
      File "/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/interfaces/pymoo/problem.py", line 4, in <module>
        from .imports import check_import, Problem, Real, Integer
    ImportError: cannot import name 'Real' from 'iwopy.interfaces.pymoo.imports' (/home/bewinche/anaconda3/envs/py310/lib/python3.10/site-packages/iwopy/interfaces/pymoo/imports.py)
    ```
  - In an environment _with_ `iwopy` installed
    ```bash
    (python-3-10-with-iwopy) user@machine:~/path/to/iwopy/repository$ python tests/core/test_discr_grid.py
    Traceback (most recent call last):
      File "/home/bewinche/modelling/reviewing/iwopy/tests/core/test_discr_grid.py", line 3, in <module>
        import iwopy
      File "/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/__init__.py", line 12, in <module>
        from . import interfaces
      File "/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/interfaces/__init__.py", line 2, in <module>
        from . import pymoo
      File "/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/interfaces/pymoo/__init__.py", line 1, in <module>
        from .optimizer import Optimizer_pymoo, DefaultCallback
      File "/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/interfaces/pymoo/optimizer.py", line 5, in <module>
        from .problem import SingleObjProblem, MultiObjProblem
      File "/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/interfaces/pymoo/problem.py", line 4, in <module>
        from .imports import check_import, Problem, Real, Integer
    ImportError: cannot import name 'Real' from 'iwopy.interfaces.pymoo.imports' (/home/bewinche/anaconda3/envs/testground/lib/python3.10/site-packages/iwopy/interfaces/pymoo/imports.py)
    ```

**Expected behavior**
From standard testing practices, a user would expect for the tests to be runnable, either with [`pytest`](https://docs.pytest.org/en/7.1.x/contents.html) or directly (e.g., `python path/to/test.py`). This should then produce a test output or some measure of success. It's possible that this issue is install related (e.g., that packages required for the testing do not come installed when installing `iwopy`as dependencies) or that documentation is missing and could be clearer around how to run these automated tests.

 - OS: Debian 11.6

**Additional context**
This was raised as part of the review for the JOSS [here](https://github.com/openjournals/joss-reviews/issues/6014).
