### 2023-11-06 FraunhoferIWES/iwopy#12 @BenWinchester
Error when importing `SimpleProblem` and `SimpleObjective`
**Describe the bug**
An `ImportError` occurs when importing the `SimpleProblem` or `SimpleObjective` classes.

**To Reproduce**
Steps to reproduce the behavior:
- Create an anaconda environment (or Python installation) initially without the `iwopy` package installed,
- Install the package:
  ```bash
  python -m pip install iwopy
  ```
- Follow the instructions in `v0.1.6` of the documentation [here](https://fraunhoferiwes.github.io/iwopy.docs/notebooks/simple_function.html) for a Simple Function minimisation;
- Observe the error:

_Traceback:_
```bash
(testground) bewinche@penguin:~/modelling/reviewing/iwopy$ python -u -m pip freeze | grep "iwopy"
iwopy==0.1.5
(testground) bewinche@penguin:~/modelling/reviewing/iwopy$ python
Python 3.10.12 (main, Jul  5 2023, 18:54:27) [GCC 11.2.0] on linux
>>> from iwopy import SimpleProblem, SimpleObjective
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/bewinche/modelling/reviewing/iwopy/iwopy/__init__.py", line 12, in <module>
    from . import interfaces
  File "/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/__init__.py", line 2, in <module>
    from . import pymoo
  File "/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/pymoo/__init__.py", line 1, in <module>
    from .optimizer import Optimizer_pymoo, DefaultCallback
  File "/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/pymoo/optimizer.py", line 5, in <module>
    from .problem import SingleObjProblem, MultiObjProblem
  File "/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/pymoo/problem.py", line 4, in <module>
    from .imports import check_import, Problem, Real, Integer
ImportError: cannot import name 'Real' from 'iwopy.interfaces.pymoo.imports' (/home/bewinche/modelling/reviewing/iwopy/iwopy/interfaces/pymoo/imports.py)
>>>
```

**Expected behavior**
A user should be able to follow the instructions without encountering an error.

 - OS: Debian 11.6

**Additional context**
This was identified as part of [the review](https://github.com/openjournals/joss-reviews/issues/6014) for the JOSS.
