### 2023-11-06 #6014 @BenWinchester
@SchmJo, I've carried out as much of the check-list review as possible but have unfortunately encountered a couple of Python `ImportError`'s and raised issued in the `iwopy` issue tracker ([here](https://github.com/FraunhoferIWES/iwopy/issues/11) and [here](https://github.com/FraunhoferIWES/iwopy/issues/12)) for these. Unfortunately, until these are resolved, I am unable to verify the functionality, performance, and automated tests.

The remained of my review is as follows below.

Ben

I think the repository and paper conform to the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and recommend it for publication. However, I have a few points, itemised below, which I would like to see addressed.

### Substantial scholarly effort
What the repository lacks in citations, it makes up for in its length of commit history and number of lines of code. The repository's nature means that, once published in the JOSS, I believe that its use will continue to expand. I am therefore happy that the repository/paper meet the requirements for substantial scholarly effort, albeit without satisfying each of the criteria.

## Checklist comments
Comments as below:

- [ ] **Contribution and authorship:**
  - [ ] Does the full list of paper authors seem appropriate and complete?
    I am unclear of the conitrbution that Bernhard Stoevesandt has made to the repository/submission. The repository's [contributors page](https://github.com/FraunhoferIWES/iwopy/graphs/contributors) shows contributions from (@SchmJo) only.

- [ ] **Functionality:**
  - [ ] Have the functional claims of the software been confirmed?
    I have encountered an error when attempting to verify by running the instructions [here](https://fraunhoferiwes.github.io/iwopy.docs/notebooks/simple_function.html). This has been raised as [issue 12](https://github.com/FraunhoferIWES/iwopy/issues/12) in the target repository.

  - [ ] See [issue 11](https://github.com/FraunhoferIWES/iwopy/issues/11) in the target issue tracker.
- [ ] **Community guidelines:**
  - [ ] Are there clear guidelines for third parties wishing to ... 3) Seek support?
    There are no clear ways to seek support through, e.g., an email mailing list, raising an issue, or opening a discussion. This isn't clear either in the [`README.md`](https://github.com/FraunhoferIWES/iwopy/blob/main/README.md) file or in the [documentation](https://fraunhoferiwes.github.io/iwopy.docs/index.html).

- [ ] **Summary:**
  - [ ] Has a clear description of the _high-level functionality_ and purpose of the software for a diverse, non-specialist audience been provided?
    The summary provides a good introduction to the problem. However, the high-level functionality is not outlined clearly enough
- [ ] **State of the field:**
  - [ ] Do the authors describe how this software compares to other commonly-used packages?
    Currently, there is no _State of the field_ section within the paper.
- [ ] **Quality of writing:**
  - [ ] Page 1, line 24, I believe that "the optimizers from the supported _liked_ packages" should read "the optimizers from the supported _*linked*_ packages." Currenly, the use of "liked" is unclear and informal. If a synonym of "liked" is wanted, potentially "preferred" should be used.
- [ ] **References:**
  For the references, the `editorialbot` has highlighted a missing DOI for `pymoo`. Other than this, I believe that the reference section should be expanded. Including the _State of the field_ section would provide scope for additional references, so I am not greatly concerned as the length of the reference list will naturally increase.
