### 2023-11-06 FraunhoferIWES/iwopy#12 @StewMH
Confirming that I see the same issue when trying to review the code.
