### 2023-11-06 FraunhoferIWES/iwopy#11 @SchmJo
Thanks for the bug report. Which version of iwopy are you running? It this the latest main branch from github, or an installation from pypi, or from conda-forge? It seems that I cannot reproduce this behaviour in my local clone of the repository. Thanks
