### 2023-11-06 FraunhoferIWES/iwopy#12 @SchmJo
Thanks for the bug report. It seems that you are mixing conda and pip installations, is that possible? The installation via conda should be done by `conda install -c conda-forge iwopy`, see the [documentation](https://fraunhoferiwes.github.io/iwopy.docs/installation.html).

My local version is a pure pip installation from the local repo clone (`pip install -e .` from the directory), and I cannot reproduce this. I will try again from the conda-forge installation.
