### 2023-11-06 FraunhoferIWES/iwopy#12 @SchmJo
Sorry, I can now indeed reprodce this for a fresh install -  for all installation approaches. I will try to figure out what is happening here.
