### 2023-11-06 FraunhoferIWES/iwopy#11 @SchmJo
I think this due to `pymoo` not being installed. For me, `pip install pymoo` solved the issue (for conda installations please install from there).

Currently the optimization packages are optional. However, this behaviour as it is now is not desired. I will think of a new solution.
