### 2024-02-01 #6014 @SchmJo
Hi all, thanks for reviewing and thanks for the issues!

I hope that I will be able to address them until mid February. As far as I see it, the issues were due to dependencies on pymoo and pygmo that were needed for the tests, but are optional for the main package. I will improve this, and go through your comments soon.

...and sorry for the delay on my side!
