### 2024-02-06 364f5fe
improved dynamic package loading
