+++ iwopy/utils/__init__.py
@@ -1,2 +1,3 @@
 from .discretization import RegularDiscretizationGrid
 from .stdout import suppress_stdout
+from .load import import_module

[…]
