+++ docs/source/history.md
@@ -122,3 +122,8 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
   - Improved dynamic imports of `pygmo` and `pymoo`
   
 **Full Changelog**: [https://github.com/FraunhoferIWES/iwopy/commits/v0.1.7](https://github.com/FraunhoferIWES/iwopy/commits/v0.1.7)
+
+## v0.1.8-beta
+
+  
+**Full Changelog**: [https://github.com/FraunhoferIWES/iwopy/commits/v0.1.8](https://github.com/FraunhoferIWES/iwopy/commits/v0.1.8)
