+++ .github/workflows/run-tests.yml
@@ -35,7 +35,6 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install .[test]
         python -m pip install pymoo
-        python -m pip install "pygmo>=2.0,<2.19"
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -45,9 +44,6 @@ jobs:
     - name: Test core
       run: |
         pytest tests/core
-    - name: Test pygmo
-      run: |
-        pytest tests/pygmo
     - name: Test pymoo
       run: |
         pytest tests/pymoo
@@ -91,7 +87,7 @@ jobs:
   tests39:
     name: run-tests-py3.9
 
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     strategy:
       fail-fast: false
       matrix:
@@ -107,7 +103,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install .[test]
-        python -m pip install pymoo
+        python -m pip install pygmo pymoo
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -117,6 +113,9 @@ jobs:
     - name: Test core
       run: |
         pytest tests/core
+    - name: Test pygmo
+      run: |
+        pytest tests/pygmo
     - name: Test pymoo
       run: |
         pytest tests/pymoo
@@ -124,7 +123,7 @@ jobs:
   tests310:
     name: run-tests-py3.10
 
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     strategy:
       fail-fast: false
       matrix:
@@ -140,7 +139,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install .[test]
-        python -m pip install pymoo
+        python -m pip install pygmo pymoo
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -150,6 +149,46 @@ jobs:
     - name: Test core
       run: |
         pytest tests/core
+    - name: Test pygmo
+      run: |
+        pytest tests/pygmo
     - name: Test pymoo
       run: |
         pytest tests/pymoo
+
+  tests311:
+    name: run-tests-py3.11
+
+    runs-on: ubuntu-22.04
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.11"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install .[test]
+        python -m pip install pygmo pymoo
+    - name: Lint with flake8
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        #flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test core
+      run: |
+        pytest tests/core
+    - name: Test pygmo
+      run: |
+        pytest tests/pygmo
+    - name: Test pymoo
+      run: |
+        pytest tests/pymoo
+        
