### 2024-02-07 7f1928a
fixing tests for python 3.7
