+++ .github/workflows/run-tests.yml
@@ -33,8 +33,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install .[test]
-        python -m pip install pymoo
+        python -m pip install pytest flake8 pymoo
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -191,4 +190,3 @@ jobs:
     - name: Test pymoo
       run: |
         pytest tests/pymoo
-        
