### 2024-02-07 5143bbd
fixing tests for python 3.7
