+++ .github/workflows/run-tests.yml
@@ -33,7 +33,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install pytest flake8 pymoo
+        python -m pip install . pytest flake8 pymoo
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
