+++ docs/source/testing.md
@@ -1,0 +1,9 @@
+# Testing
+
+For testing, please run the following steps:
+```console
+git clone https://github.com/FraunhoferIWES/iwopy.git
+cd iwopy
+pip install -e .[test]
+pytest tests
+```
