+++ docs/source/testing.md
@@ -1,9 +1,13 @@
 # Testing
 
-For testing, please run the following steps:
+For testing, please clone the repository and install the required dependencies:
 ```console
 git clone https://github.com/FraunhoferIWES/iwopy.git
 cd iwopy
 pip install -e .[test]
+```
+
+The tests are then run by
+```console
 pytest tests
 ```
