+++ docs/source/conf.py
@@ -41,14 +41,17 @@ release = __version__
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    "nbsphinx",
+    "sphinx_immaterial",
+    "sphinx_immaterial.apidoc.python.apigen",
     "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
+    "sphinx.ext.autosectionlabel",
+    # "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
-    "sphinx.ext.napoleon",
+    # "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
-    "sphinx.ext.inheritance_diagram",
-    "nbsphinx",
+    # "sphinx.ext.inheritance_diagram",
     "sphinx.ext.doctest",
     "m2r2",
 ]
@@ -60,7 +63,8 @@ intersphinx_mapping = {
 }
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
+# templates_path = ["_templates"]
+# autosummary_generate = False
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
@@ -84,16 +88,25 @@ exclude_patterns = [
     # ipynb checkpoints
     "notebooks/.ipynb_checkpoints/*.ipynb",
     "build/*",
+    # "_templates/*",
+    # DEBUG
+    # "examples.rst",
+    # "notebooks/*",
+    # "notebooks/layout_opt.ipynb",
+    # "notebooks/timelines.ipynb",
+    # "notebooks/heterogeneous.ipynb",
+    # "notebooks/timeseries.ipynb",
+    # "api.rst"
 ]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = None
 
-autosummary_generate = True
+# autosummary_generate = True
 napolean_use_rtype = False
 
 # -- Options for sphinxcontrib.email ------------------------------------------
-email_automode = True
+# email_automode = True
 
 
 # -- Options for nbsphinx -----------------------------------------------------
@@ -113,7 +126,7 @@ nbsphinx_execute = "always"
 
 
 # Controls when a cell will time out (defaults to 30; use -1 for no timeout):
-nbsphinx_timeout = 180
+nbsphinx_timeout = 500
 
 # Default Pygments lexer for syntax highlighting in code cells:
 # nbsphinx_codecell_lexer = 'ipython3'
@@ -125,14 +138,15 @@ nbsphinx_timeout = 180
 # nbsphinx_responsive_width = '700px'
 
 # This is processed by Jinja2 and inserted before each notebook
+# Fix for issue with pyplot, cf
+# https://github.com/readthedocs/sphinx_rtd_theme/issues/788#issuecomment-585785027
 nbsphinx_prolog = r"""
-{% set docname = 'docs/' + env.doc2path(env.docname, base=None) %}
-.. only:: html
-    .. role:: raw-html(raw)
-        :format: html
-    .. nbinfo::
-        :raw-html:`<a href="https://colab.research.google.com/github/DTUWindEnergy/PyWake/blob/master/{{ docname }}"><img alt="Open and run in Colab (interactive)" src="https://colab.research.google.com/assets/colab-badge.svg" style="vertical-align:text-bottom"></a>
-        <a href="https://gitlab.windenergy.dtu.dk/TOPFARM/PyWake/-/tree/master/{{ docname }}"><img alt="Edit on Gitlab" src="https://img.shields.io/badge/Edit%20on-Gitlab-blue?style=flat&logo=gitlab" style="vertical-align:text-bottom"></a>`
+.. raw:: html
+
[…]
