+++ iwopy/interfaces/pygmo/__init__.py
@@ -1,1 +1,4 @@
+from .algos import AlgoFactory
+from .imports import load
 from .optimizer import Optimizer_pygmo
+from .problem import UDP
