+++ iwopy/core/objective.py
@@ -8,7 +8,7 @@ class Objective(OptFunction):
     Abstract base class for objective functions.
 
     :group: core
-    
+
     """
 
     @abstractmethod
