+++ iwopy/__init__.py
@@ -2,6 +2,7 @@
 Fraunhofer IWES optimization tools in Python
     
 """
+
 from .core import Problem, Objective, Constraint, Memory
 from .wrappers import (
     ProblemWrapper,
