### 2024-02-08 FraunhoferIWES/iwopy#11 @SchmJo
@BenWinchester The instructions for tests have now been added to the [documentation](https://fraunhoferiwes.github.io/iwopy.docs/testing.html) and also the [README](https://github.com/FraunhoferIWES/iwopy?tab=readme-ov-file#testing).
