### 2024-02-08 FraunhoferIWES/iwopy#12 @SchmJo
Now it should show a message
```console
Module 'pymoo.core.callback' not found, maybe try 'pip install pymoo'
```
If you then follow that suggestion, it should run smoothly
