### 2024-03-14 #6014 @SchmJo
Yes, I solved the two issues that were posted by the reviewers. Waiting for feedback before closing one of them: [#12](https://github.com/FraunhoferIWES/iwopy/issues/12)
