### 2024-03-19 FraunhoferIWES/iwopy#12 @BenWinchester
Hi @SchmJo ,

It looks like this message is now popping up, telling the user that, to run the examples, they need to install `pygmo`/`pymoo`:
```python
>>> solver = Optimizer_pygmo(
...     problem,
...     problem_pars=dict(),
...     algo_pars=dict(type="ipopt", tol=1e-4),
... )
Loading pygmo
Traceback (most recent call last):
  File "/media/ben/E2A2F87FA2F8598D/Users/bened/PhD Synced/Reviewing/iwopy/iwopy/utils/load.py", line 27, in import_module
    return importlib.import_module(name, package)
  File "/usr/lib/python3.10/importlib/__init__.py", line 126, in import_module
  File "<frozen importlib._bootstrap>", line 1050, in _gcd_import
  File "<frozen importlib._bootstrap>", line 1027, in _find_and_load
  File "<frozen importlib._bootstrap>", line 1004, in _find_and_load_unlocked
ModuleNotFoundError: No module named 'pygmo'

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/media/ben/E2A2F87FA2F8598D/Users/bened/PhD Synced/Reviewing/iwopy/iwopy/interfaces/pygmo/optimizer.py", line 50, in __init__
    imports.load()
  File "/media/ben/E2A2F87FA2F8598D/Users/bened/PhD Synced/Reviewing/iwopy/iwopy/interfaces/pygmo/imports.py", line 27, in load
    pygmo = import_module("pygmo", hint="pip install pygmo")
  File "/media/ben/E2A2F87FA2F8598D/Users/bened/PhD Synced/Reviewing/iwopy/iwopy/utils/load.py", line 31, in import_module
    raise ModuleNotFoundError(f"Module '{mdl}' not found, maybe try '{hint}'")
ModuleNotFoundError: Module 'pygmo' not found, maybe try 'pip install pygmo'
```

In the review for the Journal of Open-Source Software, there's a requirement which is
> [ ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Whilst this pop-up to the user does work, I think it would be good to have, perhaps in the `README.md` or even in the test documentation, a list of dependencies that need to be installed in order for a user to run the package. As it is, the automated dependencies that come with `iwopy` don't let you run the tests, and they're not really documented (other than in this issue as far as I can tell).
