### 2024-03-19 #6014 @BenWinchester
Hi @SchmJo and @fraukewiese ,

Following the advice within [issue #12](https://github.com/FraunhoferIWES/iwopy/issues/12) of the `iwopy` repository, I installed the dependencies and can confirm that the software runs as described. However, whilst this means that I have crossed off the functionality-related requirements, there are still
- the other outstanding comments (which are hopefully visible in the updated review and checklist above),
- and a new comment (indicated with :sparkles: and the date, **19/3/24**) in that, whilst a user can manually install the previously-missing dependencies, there is no clear documentation of these dependencies, either in the user guide, `README.md`, repository folder etc., which means that I don't believe, in its current form, the repository contains "a clearly-stated list of dependencies." Any of the above methods (`README.md`, inclusion in the `requirements.txt`, etc.) would all suffice, so I believe this isn't a major issue, but thought I'd flag so it's kept track of :smiley:
