### 2024-03-20 #6014 @SchmJo
@BenWinchester Concerning the deps:
- The above discussion is about a dependency that is required for running the tests. It appears [here](https://github.com/FraunhoferIWES/iwopy/blob/fcde3d7004a02d53889f71d646cf060b0f0e98df/setup.cfg#L39) in the `setup.cfg` file
- The testing is described [here](https://github.com/FraunhoferIWES/iwopy/tree/main?tab=readme-ov-file#testing) in the `README.md` and [here](https://fraunhoferiwes.github.io/iwopy.docs/testing.html) in the documentation
- Both descriptions explicitly state the instruction `pip install -e .[test]`, which installs these dependencies

Is this insufficient? If you insist, I can make it a general dependency, even though this somewhat breaks the intention of iwopy to not install all packages automatically for which it provides interfaces. Thanks for further feedback
