### 2024-03-20 #6014 @BenWinchester
@SchmJo , thanks for getting back to me :smile:

I think there isn't really a best way to deal with this dependency. I may be interpreting this wrong (sorry if so!) but it seems as though, `iwopy`, as a package, works to wrap around other optimisation tools, so it's dependent on them. If it's nicer to have it not install these dependencies (so users can choose their own optimisation packages to install?), then they probably don't need to be automatic. However, there's the check-list comment that needs addressing:
> [ ] Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I think, at the minute, listing the dependencies in the `setup.cfg` as installed only for testing isn't sufficient for a user to run the code. _E.G._, for someone to run the examples (as I did: they're great fun!), they need to run the code, get the error message, realise that they should install the packages manually, and then all will be well.

This doesn't seem neat enough to me to fulfil the "clearly-stated list" requirement. They could maybe go in the `README.md`, or, if that breaks the idea that the packages can be selected independently from `iwopy`, perhaps they could go at the top of each example in the documentation? Something like "to run these examples, you'll need to have `pygmo` installed. You can run `pip install pygmo`" or something similar? There's nothing in the checkbox (as far as I can read) that says the process needs to be automatic, just that it needs to be clearly-stated :smiley:
