### 2024-04-12 #6014 @SchmJo
I am not sure I understand the issue with the not explicitly mentioned dependencies - it is a `setup.cfg`-type package. For me it is not nice to explicitly state the dependencies again as `pip install ...` at several places, since this is redundant and harder to maintain. What is your opinion, @fraukewiese ? Is `pip install .[test]` really insufficient for the user, before running tests?
