### 2024-04-12 #6014 @BenWinchester
@SchmJo , happy to leave this to @fraukewiese . Just to be clear though, my issue is that, to run the code, _i.e._, to run any of the examples in the documentation (which are great fun!), you need these dependencies installed, and there's currently no guidance for this: a user only finds out when the error message is displayed to them. One of the options I mentioned, which doesn't seem to tricky to implement (as it just requires an update to the source code behind the examples documentation—no changes are needed to the source code of the package), is having a message at the top of the examples with some short instructions:
> Something like "to run these examples, you'll need to have pygmo installed. You can run pip install pygmo" or something similar? There's nothing in the checkbox (as far as I can read) that says the process needs to be automatic, just that it needs to be clearly-stated 😃
