+++ docs/source/installation.md
@@ -52,6 +52,11 @@ This in general corresponds to the `main` branch at [github](https://github.com/
 pip install git+https://github.com/FraunhoferIWES/iwopy@dev#egg=iwopy
 ```
 
+Notice that the above default installation does not install the third-party optimization
+packages. `iwopy` will tell you in an error message that it is missing a package, with
+a hint of installation advice. You can avoid this step by installing all supported
+optimzer packages by repla
+
 ### Developers
 
 The first step as a developer is to clone the `iwopy` repository by
