+++ setup.cfg
@@ -36,6 +36,9 @@ install_requires =
     matplotlib
 
 [options.extras_require]
+opt = 
+    pygmo
+    pymoo>=0.6
 test =
     flake8
     pytest
