+++ requirements.txt
@@ -1,38 +1,33 @@
-#
-# This file is autogenerated by pip-compile with Python 3.8
-# by the following command:
-#
-#    pip-compile --output-file=requirements.txt
-#
-contourpy==1.1.1
-    # via matplotlib
+about-time==4.2.1
+alive-progress==3.1.5
+autograd==1.6.2
+cloudpickle==3.0.0
+cma==3.2.2
+contourpy==1.2.1
 cycler==0.12.1
-    # via matplotlib
-fonttools==4.48.1
-    # via matplotlib
-importlib-resources==6.1.1
-    # via matplotlib
+Deprecated==1.2.14
+dill==0.3.8
+exceptiongroup==1.2.1
+flake8==7.0.0
+fonttools==4.52.4
+future==1.0.0
+grapheme==0.6.0
+iniconfig==2.0.0
 kiwisolver==1.4.5
-    # via matplotlib
-matplotlib==3.7.4
-    # via iwopy (setup.py)
-numpy==1.24.4
-    # via
-    #   contourpy
-    #   iwopy (setup.py)
-    #   matplotlib
-    #   scipy
-packaging==23.2
-    # via matplotlib
-pillow==10.2.0
-    # via matplotlib
-pyparsing==3.1.1
-    # via matplotlib
-python-dateutil==2.8.2
-    # via matplotlib
-scipy==1.10.1
-    # via iwopy (setup.py)
+matplotlib==3.9.0
+mccabe==0.7.0
+numpy==1.26.4
+packaging==24.0
+pillow==10.3.0
+pluggy==1.5.0
+pycodestyle==2.11.1
+pyflakes==3.2.0
+pygmo==2.19.5
+pymoo==0.6.1.1
+pyparsing==3.1.2
+pytest==8.2.1
+python-dateutil==2.9.0.post0
+scipy==1.13.1
 six==1.16.0
-    # via python-dateutil
-zipp==3.17.0
-    # via importlib-resources
+tomli==2.0.1
+wrapt==1.16.0
