+++ setup.cfg
@@ -51,6 +51,7 @@ doc =
     ipykernel
     ipywidgets
     m2r2
+    lxml_html_clean
 all =
     flake8
     pytest
