### 2024-05-30 FraunhoferIWES/iwopy#12 @SchmJo
@BenWinchester can we close this issue, considering the recent additions here: https://github.com/openjournals/joss-reviews/issues/6014#issuecomment-2137820074
