### 2024-06-05 FraunhoferIWES/iwopy#12 @BenWinchester
@SchmJo , yes! Feel free to close to the issue :smiley:
(Apologies for the delay, I'm currently travelling with limited access to a reliable and fast internet connection.)
