### 2024-06-12 #6014 @fraukewiese
@SchmJo : I think @BenWinchester has described the conditions for being mentioned as a co-author very well. So I think it is on your and your potential co-authors side to decide whether Bernhard Stoevesandt has contributed more than purely financial or organisational contributions.
