+++ requirements.txt
@@ -7,27 +7,18 @@ contourpy==1.2.1
 cycler==0.12.1
 Deprecated==1.2.14
 dill==0.3.8
-exceptiongroup==1.2.1
-flake8==7.0.0
-fonttools==4.52.4
+fonttools==4.53.0
 future==1.0.0
 grapheme==0.6.0
-iniconfig==2.0.0
 kiwisolver==1.4.5
 matplotlib==3.9.0
-mccabe==0.7.0
 numpy==1.26.4
-packaging==24.0
-pillow==10.3.0
-pluggy==1.5.0
-pycodestyle==2.11.1
-pyflakes==3.2.0
+packaging==24.1
+pillow==10.4.0
 pygmo==2.19.5
 pymoo==0.6.1.1
 pyparsing==3.1.2
-pytest==8.2.1
 python-dateutil==2.9.0.post0
-scipy==1.13.1
+scipy==1.14.0
 six==1.16.0
-tomli==2.0.1
 wrapt==1.16.0
