+++ setup.cfg
@@ -17,7 +17,15 @@ project_urls =
     Documentation = https://fraunhoferiwes.github.io/iwopy.docs/index.html
 
 classifiers =
+    Topic :: Scientific/Engineering
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
     Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Development Status :: 4 - Beta
@@ -28,10 +36,10 @@ include_package_data = True
 package_dir =
 packages = find:
 python_requires = 
-    >=3.7
+    >=3.8
 
 install_requires =
-    numpy
+    numpy<2
     scipy
     matplotlib
 
