+++ setup.cfg
@@ -21,6 +21,7 @@ classifiers =
     Intended Audience :: Developers
     Intended Audience :: Science/Research
     Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
@@ -36,7 +37,7 @@ include_package_data = True
 package_dir =
 packages = find:
 python_requires = 
-    >=3.8
+    >=3.7
 
 install_requires =
     numpy<2
