+++ setup.cfg
@@ -2,7 +2,7 @@
 name         = iwopy 
 version      = file: iwopy/VERSION
 author       = Fraunhofer IWES
-author_email = jonas.schmidt@iwes.fraunhofer.de
+author_email = jonas.schulte@iwes.fraunhofer.de
 description  = Fraunhofer IWES optimization tools in Python
 keywords     = Optimization, Interface
 license      = MIT
