### 2024-07-04 #6014 @SchmJo
@BenWinchester I just pushed updates to the `paper` branch, hopefully addressing all your doubts:

### Authorship
Bernhard Stoevesandt has been removed as an author.

### Paper summary
_The summary provides a good introduction to the problem. However, the high-level functionality is not outlined clearly enough_

In my opinion the sections "Summary" and "Statement of need" together provide detailed insight in the technicalities behind `iwopy`. The sentence below was added at the end of the "Summary" section, hoping that this addresses your concerns:

```
The package `iwopy` solves this problem by providing a convenient
and flexible interface which is general enough to be applicable to a
wide range of optimization problems.
```

### State of the field

_Currently, there is no State of the field section within the paper._

To my understanding there should only be two sections in a JOSS paper, namely "Summary" and "Statement of need". The following paragraph was inserted into the "Statement of need", stressing the possibilities of existing optimization packages:

```
Note that more optimization packages are available which are not yet supported,
like `pyomo [@pyomo]`, `Platypus [@platypus]`, `DEAP [@deap]` and others.
Each package is well suited for solving a wide range of optimization problems, and they
all come with extensive user interfaces. However, `iwopy`
addresses a unification of those interfaces, enabling the user to benefit from all supported
optimizers without the need of extensive changes of the code base.
```

### Quality of writing

_Page 1, line 24, I believe that "the optimizers from the supported liked packages" should read "the optimizers from the supported linked packages." Currenly, the use of "liked" is unclear and informal. If a synonym of "liked" is wanted, potentially "preferred" should be used._

Thanks, the typo was corrected. Indeed it was supposed to read "linked".

### References

_For the references, the editorialbot has highlighted a missing DOI for pymoo. Other than this, I believe that the reference section should be expanded. Including the State of the field section would provide scope for additional references, so I am not greatly concerned as the length of the reference list will naturally increase._

Four more references were added: scipy, pyomo, Platypus and DEAP. I hope this is sufficient.

I hope that all concers have been addressed by this? Thanks again for reviewing!
