### 2024-08-07 a99a96a
adding support advice to docs
