### 2024-08-07 a032a46
adding support advice to docs
