### 2024-08-07 #6014 @SchmJo
@BenWinchester I added a section on how to seek support:
- [here](https://github.com/FraunhoferIWES/iwopy?tab=readme-ov-file#support) in the README.md
- and [here](https://fraunhoferiwes.github.io/iwopy.docs/index.html#support) in the documentation

I hope this is what you had in mind? Thanks
