### 2024-08-23 75e75e7
recovering numpy >= 2
