+++ conda/meta.yaml
@@ -1,0 +1,49 @@
+{% set data = load_setup_py_data() %}
+
+package:
+  name: iwopy
+  version: {{ data['version'] }}
+
+source:
+  path: ..
+
+build:
+  number: 0
+  script: python setup.py install --single-version-externally-managed --record=record.txt
+
+requirements:
+  build:
+    - python
+    - matplotlib-base
+    - numpy
+    - scipy
+  run:
+    - python
+    - matplotlib-base
+    - numpy
+    - scipy
+
+test:
+  imports:
+    - iwopy
+  requires:
+    - pip
+    - pytest
+    - pygmo
+    - pymoo
+  source_files:
+    - tests
+  commands:
+    - pip check
+    - pytest tests
+
+about:
+  home: {{ data['project_urls']['Source Code'] }}
+  license: {{ data['license'] }}
+  summary: {{ data['description'] }}
+  description: |
+    The software iwopy is a modular wind farm simulation and wake
+    modelling toolbox which is based on engineering wake models. It
+    has many applications, for example wind farm optimization,
+    wind farm post-construction analysis, wake model studies, and
+    wind farm simulations invoking complex model chains.
