+++ setup.cfg
@@ -1,76 +1,0 @@
-[metadata]
-name         = iwopy 
-version      = file: iwopy/VERSION
-author       = Fraunhofer IWES
-author_email = jonas.schulte@iwes.fraunhofer.de
-description  = Fraunhofer IWES optimization tools in Python
-keywords     = Optimization, Interface
-license      = MIT
-
-long_description = file: README.md
-long_description_content_type = text/markdown
-
-url = https://github.com/FraunhoferIWES/iwopy
-project_urls =
-    Source Code = https://github.com/FraunhoferIWES/iwopy
-    Bug Tracker = https://github.com/FraunhoferIWES/iwopy/issues
-    Documentation = https://fraunhoferiwes.github.io/iwopy.docs/index.html
-
-classifiers =
-    Topic :: Scientific/Engineering
-    Intended Audience :: Developers
-    Intended Audience :: Science/Research
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-    Development Status :: 4 - Beta
-
-[options]
-zip_safe = True
-include_package_data = True
-package_dir =
-packages = find:
-python_requires = 
-    >=3.7
-
-install_requires =
-    numpy<2
-    scipy
-    matplotlib
-
-[options.extras_require]
-opt = 
-    pygmo
-    pymoo>=0.6
-test =
-    flake8
-    pytest
-    pygmo
-    pymoo>=0.6
-doc = 
-    sphinx 
-    sphinx-immaterial
-    nbsphinx
-    ipykernel
-    ipywidgets
-    m2r2
-    lxml_html_clean
-all =
-    flake8
-    pytest
-    pygmo
-    pymoo>=0.6
-    sphinx 
-    sphinx-immaterial
-    nbsphinx
-    ipykernel
-    ipywidgets
-    m2r2
-    
-scripts =
