+++ conda/meta.yaml
@@ -42,8 +42,6 @@ about:
   license: {{ data['license'] }}
   summary: {{ data['description'] }}
   description: |
-    The software iwopy is a modular wind farm simulation and wake
-    modelling toolbox which is based on engineering wake models. It
-    has many applications, for example wind farm optimization,
-    wind farm post-construction analysis, wake model studies, and
-    wind farm simulations invoking complex model chains.
+    The iwopy package is a meta package that provides interfaces 
+    to other open-source Python optimization packages out there.
+    
