### 2024-09-05 #6014 @fraukewiese
@SchmJo : Could you please add References for the four different packages you mention in the paper part of the submission? For pymoo and pygmo, you do, but for scipy, pyomo, platypus and deap the reference is missing.
