+++ tests/pygmo/test_pygmo.py
@@ -1,415 +1,0 @@
-import numpy as np
-
-from iwopy import DiscretizeRegGrid, SimpleConstraint
-from iwopy.benchmarks.branin import BraninProblem
-from iwopy.benchmarks.rosenbrock import RosenbrockProblem
-from iwopy.interfaces.pygmo import Optimizer_pygmo
-
-
-class RC(SimpleConstraint):
-    def __init__(self, problem, name="c", ana_deriv=False):
-        super().__init__(problem, name, n_components=2, has_ana_derivs=ana_deriv)
-
-    def f(self, x, y):
-        return [(x - 1) ** 3 - y + 1, x + y - 3]
-
-    def g(self, var, x, y, components):
-
-        out = np.full(len(components), np.nan, dtype=np.float64)
-        for i, ci in enumerate(components):
-
-            # (x-1)**3 - y + 1
-            if ci == 0:
-                out[i] = 3 * (x - 1) ** 2 if var == 0 else -1
-
-            # x + y - 3
-            elif ci == 1:
-                out[i] = 1
-
-        return out
-
-
-def run_branin_ipopt(init_vals, dx, dy, tol, pop):
-
-    prob = BraninProblem(initial_values=init_vals)
-    gprob = DiscretizeRegGrid(prob, deltas={"x": dx, "y": dy}, interpolation="linear")
-    gprob.initialize()
-
-    solver = Optimizer_pygmo(
-        gprob,
-        problem_pars=dict(pop=pop),
-        algo_pars=dict(type="ipopt", tol=tol),
-    )
-    solver.initialize()
-
-    results = solver.solve()
-    solver.finalize(results)
-
-    return results
-
-
-def test_branin_ipopt():
-
-    cases = (
-        (
-            0.001,
-            0.001,
-            1e-6,
-            (1.0, 1.0),
-            0.397887,
-            (9.42478, 2.475),
-            5e-6,
-            (0.0008, 0.002),
-            False,
-        ),
-        (
-            0.0001,
-            0.0001,
-            1e-7,
-            (1.0, 1.0),
-            0.397887,
-            (9.42478, 2.475),
-            5e-7,
-            (7e-5, 1.1e-4),
-            False,
-        ),
-        (
-            0.001,
-            0.001,
-            1e-6,
-            (-3, 12.0),
-            0.397887,
-            (-np.pi, 12.275),
-            7e-6,
-            (0.001, 0.0016),
-            True,
-        ),
-        (
-            0.001,
-            0.001,
-            1e-6,
-            (3, 3.0),
-            0.397887,
-            (np.pi, 2.275),
-            3e-6,
-            (0.0005, 0.00015),
-            True,
-        ),
-    )
-
-    for dx, dy, tol, ivals, f, xy, limf, limxy, pop in cases:
-
-        print("\nENTERING", (dx, dy, tol, ivals, f, xy, limf, limxy, pop), "\n")
-
-        results = run_branin_ipopt(ivals, dx, dy, tol, pop)
-        print("Opt vars:", results.vars_float)
-
-        delf = np.abs(results.objs[0] - f)
-        print("delf =", delf, ", lim =", limf)
-        assert delf < limf
-
-        delxy = np.abs(results.vars_float - np.array(xy))
-        limxy = np.array(limxy)
-        print("delxy =", delxy, ", lim =", limxy)
-        assert np.all(delxy < limxy)
-
-
-def run_branin_heu(type, init_vals, ngen, npop):
-
-    prob = BraninProblem(initial_values=init_vals)
-    prob.initialize()
-
-    solver = Optimizer_pygmo(
-        prob,
-        problem_pars=dict(),
-        algo_pars=dict(
-            type=type,
[…]
