### 2024-09-10 a4e69f7
adding tests for Python3.12
