+++ pyproject.toml
@@ -1,8 +1,8 @@
 # pyproject.toml file specified at the root of the directory
 
 [build-system]
-requires = ["pdm-backend"]
-build-backend = "pdm.backend"
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
 
 [project]
 name = "iwopy"
