### 2024-09-10 7915a8b
dropping support for Python 3.7
