+++ .github/workflows/run-tests.yml
@@ -15,38 +15,6 @@ on:
       - 'dev'
 
 jobs:
-  tests37:
-    name: run-tests-py3.7
-
-    runs-on: ubuntu-20.04
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.7"]
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install . pytest flake8 pymoo
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        #flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test core
-      run: |
-        pytest tests/core
-    - name: Test pymoo
-      run: |
-        pytest tests/pymoo
-
   tests38:
     name: run-tests-py3.8
 
@@ -211,4 +179,3 @@ jobs:
     - name: Test pymoo
       run: |
         pytest tests/pymoo
-        
