+++ requirements.txt
@@ -1,24 +1,30 @@
 about-time==4.2.1
 alive-progress==3.1.5
-autograd==1.6.2
-cloudpickle==3.0.0
+autograd==1.7.0
 cma==3.2.2
 contourpy==1.2.1
 cycler==0.12.1
 Deprecated==1.2.14
 dill==0.3.8
-fonttools==4.53.0
-future==1.0.0
+flake8==7.1.1
+fonttools==4.53.1
 grapheme==0.6.0
+iniconfig==2.0.0
 kiwisolver==1.4.5
-matplotlib==3.9.0
-numpy==1.26.4
+matplotlib==3.9.2
+mccabe==0.7.0
+numpy==2.1.0
 packaging==24.1
 pillow==10.4.0
-pygmo==2.19.5
-pymoo==0.6.1.1
+pluggy==1.5.0
+pycodestyle==2.12.1
+pyflakes==3.2.0
+pymoo==0.6.1.3
 pyparsing==3.1.2
+pytest==8.3.2
 python-dateutil==2.9.0.post0
-scipy==1.14.0
+scipy==1.14.1
+setuptools==73.0.1
 six==1.16.0
+wheel==0.44.0
 wrapt==1.16.0
