+++ paper/paper.bib
@@ -46,9 +46,14 @@
   doi     = {10.1038/s41592-019-0686-2},
 }
 
-@book{pyomo, title={Pyomo--optimization modeling in python}, 
+@book{pyomo, 
+title={Pyomo--optimization modeling in python}, 
 author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Laird, Carl D. and Nicholson, Bethany L. and Siirola, John D. and Watson, Jean-Paul and Woodruff, David L.}, 
-edition={Third}, volume={67}, year={2021}, publisher={Springer Science \& Business Media} 
+edition={Third}, 
+volume={67}, 
+year={2021}, 
+publisher={Springer Science \& Business Media},
+doi={10.1007/978-3-030-68928-5},
 } 
 
 @Manual{platypus,
@@ -56,6 +61,7 @@ edition={Third}, volume={67}, year={2021}, publisher={Springer Science \& Busine
   author = {Alexander Yermanos},
   year = {2023},
   note = {R package version 3.5.0},
+  doi = {10.1093/nargab/lqab023},
 }
 
 @article{deap,
@@ -66,4 +72,5 @@ edition={Third}, volume={67}, year={2021}, publisher={Springer Science \& Busine
     month     = { jul },
     year      = { 2012 },
     journal   = { Journal of Machine Learning Research }
+    doi       = {10.5555/2503308.2503311},
 }
