+++ paper/paper.bib
@@ -56,14 +56,6 @@ publisher={Springer Science \& Business Media},
 doi={10.1007/978-3-030-68928-5},
 } 
 
-@Manual{platypus,
-  title = {Platypus: Single-Cell Immune Repertoire and Gene Expression Analysis},
-  author = {Alexander Yermanos},
-  year = {2023},
-  note = {R package version 3.5.0},
-  doi = {10.1093/nargab/lqab023},
-}
-
 @article{deap,
     author    = " F\'elix-Antoine Fortin and Fran\c{c}ois-Michel {De Rainville} and Marc-Andr\'e Gardner and Marc Parizeau and Christian Gagn\'e ",
     title     = { {DEAP}: Evolutionary Algorithms Made Easy },
@@ -73,4 +65,12 @@ doi={10.1007/978-3-030-68928-5},
     year      = { 2012 },
     journal   = { Journal of Machine Learning Research },
     doi       = {10.5555/2503308.2503311},
+}
+
+@Manual{platypus,
+  title = {Platypus: Single-Cell Immune Repertoire and Gene Expression Analysis},
+  author = {Alexander Yermanos},
+  year = {2023},
+  note = {R package version 3.5.0},
+  doi = {10.1093/nargab/lqab023},
 }
