+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "iwopy"
-version = "0.3"
+version = "0.3.1"
 authors = [
   {name = "Jonas Schulte"},
   {name = "Bernhard Stoevesandt"},
