+++ .github/workflows/pypi.yml
@@ -68,7 +68,7 @@ jobs:
         name: python-package-distributions
         path: dist/
     - name: Sign the dists with Sigstore
-      uses: sigstore/gh-action-sigstore-python@master
+      uses: sigstore/gh-action-sigstore-python@main
       with:
         inputs: >-
           ./dist/*.tar.gz
