### 2024-09-10 7bdae4e
removing broken doi
