+++ paper/paper.bib
@@ -64,7 +64,6 @@ doi={10.1007/978-3-030-68928-5},
     month     = { jul },
     year      = { 2012 },
     journal   = { Journal of Machine Learning Research },
-    doi       = {10.5555/2503308.2503311},
 }
 
 @Manual{platypus,
