### 2024-09-10 #6014 @SchmJo
- I removed the broken DOI, but I did not manage to find a better one
- I think Zenodo shows the correct title, does it not?

![image](https://github.com/user-attachments/assets/d65b89c3-f1aa-4160-bce3-f218a064b74d)
