### 2024-09-10 #6014 @SchmJo
I just pushed another fix for the broken doi, as recommended by the bot: [https://dl.acm.org/doi/10.5555/2503308.2503311](https://dl.acm.org/doi/10.5555/2503308.2503311)
