### 2024-10-02 #6014 @SchmJo
Oh, sorry. Here is the new data:
- DOI: 10.5281/zenodo.13739797
- Zenodo: https://doi.org/10.5281/zenodo.13739797
