### 2024-10-07 #6014 @crvernon
:wave: @SchmJo - I just need you to address the following before I can move forward with accepting this for publication:

**In your archive:**
- [ ] In your Zenodo archive, please edit the license in the metadata to match the one you are using in your repository.

**In your paper:**
- [ ] In the references section of your paper, please ensure that capitalization is maintained where necessary.  For example in LINE 49, the "p" in "python" should be capitalized.  Please look through your references to ensure that others have also been addressed.  Formatting can be maintained in your bib file using curly brackets around the characters you wish to maintain formatting of.

A new release is not required for updating the above.  Let me know when you finish and I'll take a look.
