+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{pymoo,
-  title={Pymoo: Multi-Objective Optimization in Python},
+  title={{Pymoo: Multi-Objective Optimization in Python}},
   author={Julian Blank and Kalyanmoy Deb},
   journal={IEEE Access},
   year={2020},
@@ -47,7 +47,7 @@
 }
 
 @book{pyomo, 
-title={Pyomo--optimization modeling in python}, 
+title={Pyomo--optimization modeling in {P}ython}, 
 author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Laird, Carl D. and Nicholson, Bethany L. and Siirola, John D. and Watson, Jean-Paul and Woodruff, David L.}, 
 edition={Third}, 
 volume={67}, 
