### 2024-10-09 0175344
remove duplicate brackets
