### 2024-10-09 #6014 @crvernon
:wave: @SchmJo - there was still a capitalization issue in your references, so I set up a PR from a fork into your `paper` branch that should take care of this.  Please see and review.  Let me know when this makes it into your branch and I'll check the status again.  Thanks!
