### 2024-10-09 b2f1851
JOSS Review:  Capitalization fix (#13)
* fix capitalization

* remove duplicate brackets
