+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{pymoo,
-  title={Pymoo: Multi-Objective Optimization in Python},
+  title={{Pymoo: Multi-Objective Optimization in Python}},
   author={Julian Blank and Kalyanmoy Deb},
   journal={IEEE Access},
   year={2020},
@@ -36,7 +36,7 @@
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+  title   = {{SciPy 1.0: Fundamental Algorithms for Scientific
             Computing in Python}},
   journal = {Nature Methods},
   year    = {2020},
@@ -47,7 +47,7 @@
 }
 
 @book{pyomo, 
-title={Pyomo--optimization modeling in python}, 
+title={Pyomo--optimization modeling in {P}ython}, 
 author={Bynum, Michael L. and Hackebeil, Gabriel A. and Hart, William E. and Laird, Carl D. and Nicholson, Bethany L. and Siirola, John D. and Watson, Jean-Paul and Woodruff, David L.}, 
 edition={Third}, 
 volume={67}, 
