### 2024-10-09 #6014 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
